backend/pyproject.toml
[project]
name = "lightrag-backend"
version = "0.1.0"
description = "Conversational knowledge graph backend powered by LightRAG"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "lightrag-hku>=1.4.9.3",
    "networkx>=3.2.1",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "numpy",
    "httpx>=0.25.0",
    "cryptography>=41.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "openai~=1.46",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
